Convert to rollup build & remove 'brfs' dependency
This removes the need for packages / applications consuming this
package in the browser to use the brfs browserify transform to
handle the fact that this package uses node's 'fs' module.

The bundled code has brfs applied at build time to inline the file
contents that must be read using 'fs'.